+++ src/App.tsx
@@ -137,7 +137,7 @@ const components: DemoComponent[] = [
   },
   {
     name: "patient-adt-table",
-    render: () => <PatientsADTAlertsTable />,
+    render: () => <PatientsADTAlertsTable data={[]} />,
     title: "Patient ADT Alerts Table",
   },