+++ src/components/content/medications/helpers/add-new-med-drawer.tsx
@@ -16,7 +16,7 @@ export function useAddMedicationForm() {
 
   return (medication: MedicationStatementModel) => {
     openDrawer({
-      telemetryName: "add_medication",
+      trackingMetadata: { action: "add_medication" },
       component: (props) => <AddNewMedDrawer medication={medication.resource} {...props} />,
     });
   };